refactor: split server stats viewer routes
